move test location, add some `TYPE_CHECKING_INVALID_USAGE`
