+++ tests/series/test_indexing.py
@@ -8,6 +8,7 @@ from typing_extensions import assert_type
 
 from tests import (
     PD_LTE_23,
+    TYPE_CHECKING_INVALID_USAGE,
     check,
     pytest_warns_bounded,
 )
@@ -35,6 +36,9 @@ def test_types_iloc_iat() -> None:
     s2.loc[0]
     s2.iat[0]
     s2.iat[0] = None
+
+    if TYPE_CHECKING_INVALID_USAGE:
+        s.iat[0, 0]  # type: ignore[index]  # pyright: ignore[reportArgumentType]
 
 
 def test_types_loc_at() -> None:
